requirements.txt
torch
torchvision
pillow
requests
zmq
pytest